feat(kratos): support specifying deployment env vars from a secret (#190)
When setting up Kratos via Helm there are currently secrets in the config that should be passed in via another means.
The case I'm dealing with specifically is kratos.config.selfservice.methods.oidc.config.providers, which contains client_id and client_secret for each provider.
In this case my ideal scenario seems to be to set up secret containing the provider config, and then have that create env vars which Kratos can use to override config (in this case: SELFSERVICE_METHODS_OIDC.)
This requires that the secret is created separately, but I think that is beneficial. If it were up to Helm to create the secret then anyone who needed to run the chart would also have to have the contents of the secret. And in my case I'm executing Helm with Terraform, which would require the secrets to go into the Terraform state, and be accessible by anyone who needs to run that TF.