feat: add watcher pilot (#293)
This PR introduces a sidecar container to each deployment using the DSN mechanism. It allows setting a watcher for a mounted file, and restarting the deployment upon changes in the file. 

This uses inotify-tools, which report upon file changes. In this case we are operating with k8s volumes, which are mounted in some location and symlinked into the desired location. Change to the original file (so unmount and remount) causes the symlink to be deleted and recreated. In some cases this can break the program (f.e if the mounted file is a secret/certificate used by DSN) requiring a restart.