expose --endpoint argument in chart
Hydra 2.x uses /admin/clients, Hydra 1.x uses /clients.

Since hydra 2.x returns a 307, this will likely not work inside an
environment like kubernetes where the request is made to
http://my-svc:4445.

The 307 will return the destination as (from curl):

    < Location: https://localhost:4445/admin/clients

This causes maester to go into a loop where it crashes on the
redirect.